Fixed incorrect data type
Former-commit-id: a33e0a21fc5b4db947bbedce44b3bdf45f8a9f0b

+++ exporters/FusionExporter/Source/Data/BXDJ/Components/Wheel.cpp
@@ -135,7 +135,7 @@ rapidjson::Value Wheel::getJSONObject(rapidjson::MemoryPoolAllocator<>& allocato
 
 	wheelJSON.AddMember("type", rapidjson::Value((int)type), allocator);
 	wheelJSON.AddMember("frictionLevel", rapidjson::Value((int)frictionLevel), allocator);
-	wheelJSON.AddMember("isDriveWheel", rapidjson::Value((int)isDriveWheel), allocator);
+	wheelJSON.AddMember("isDriveWheel", rapidjson::Value(isDriveWheel), allocator);
 
 	return wheelJSON;
 }
